requirements.txt
# Core dependencies
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.9.0,<2.0.0
pandas>=1.5.0,<2.0.0
openpyxl>=3.1.0,<4.0.0  # For Excel support in pandas
pyyaml>=6.0,<7.0
psycopg2-binary>=2.9.5,<3.0.0  # PostgreSQL driver
python-dateutil>=2.8.2,<3.0.0  # Date parsing utilities
freezegun>=1.2.0,<2.0.0  # For time-based testing

# Development dependencies
pytest>=7.0.0,<8.0.0
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.0.0,<2.0.0
ruff>=0.0.262,<1.0.0

# Type checking
types-PyYAML>=6.0.0,<7.0.0
types-requests>=2.29.0,<3.0.0
types-setuptools>=65.0.0,<66.0.0
types-toml>=0.10.0,<0.11.0

# Logging
loguru>=0.6.0,<0.7.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
coverage>=7.0.0,<8.0.0
